+++ core/helpers.py
@@ -7,3 +7,11 @@ def create_user_profile(sso_session_id, data):
     )
     response.raise_for_status()
     return response
+
+
+def update_user_profile(sso_session_id, data):
+    response = sso_api_client.user.update_user_profile(
+        sso_session_id=sso_session_id, data=data
+    )
+    response.raise_for_status()
+    return response