+++ profile/fab/tests/test_views.py
@@ -231,29 +231,10 @@ def test_supplier_company_retrieve_found_business_profile_on(
     assert response.template_name == ['fab/profile.html']
 
 
-def test_company_owner(client, user):
+@pytest.mark.parametrize('param', ('owner-transferred', 'user-added', 'user-removed'))
+def test_success_message(mock_retrieve_supplier, client, param, user):
     client.force_login(user)
-    response = client.get(reverse('find-a-buyer'))
-
-    assert response.context_data['is_profile_owner'] is True
-
-
-def test_non_company_owner(mock_retrieve_supplier, client, user):
-    client.force_login(user)
-    mock_retrieve_supplier.return_value = create_response({'is_company_owner': False})
-    response = client.get(reverse('find-a-buyer'))
-
-    assert response.context_data['is_profile_owner'] is False
-
-
-@pytest.mark.parametrize('param', (
-    'owner-transferred', 'user-added', 'user-removed'
-))
-def test_success_message(
-    mock_retrieve_supplier, client, param, user
-):
-    client.force_login(user)
-    mock_retrieve_supplier.return_value = create_response({'is_company_owner': False})
+    mock_retrieve_supplier.return_value = create_response({'role': user_roles.EDITOR})
 
     url = reverse('find-a-buyer')
     response = client.get(url, {param: True})
@@ -801,6 +782,42 @@ def test_personal_details(client, mock_create_user_profile, user):
             'mobile_phone_number': '1232342'
         }
     )
+
+
+def test_personal_profile_edit(client, mock_update_user_profile, user):
+    client.force_login(user)
+    data = {
+        'given_name': 'Foo',
+        'family_name': 'Example',
+        'job_title': 'Exampler',
+        'phone_number': '1232342',
+    }
+
+    response = client.get(reverse('find-a-buyer-personal-profile:edit'))
+
+    assert response.status_code == 200
+
+    response = client.post(reverse('find-a-buyer-personal-profile:edit'), data)
+
+    assert response.status_code == 302
+    assert response.url == reverse('find-a-buyer-personal-profile:display')
+    assert mock_update_user_profile.call_count == 1
+    assert mock_update_user_profile.call_args == mock.call(
+        sso_session_id=user.session_id,
+        data={
+            'first_name': 'Foo',
+            'last_name': 'Example',
+            'job_title': 'Exampler',
+            'mobile_phone_number': '1232342'
+        }
+    )
+
+
+def test_personal_profile_display(client, user):
+    client.force_login(user)
+    response = client.get(reverse('find-a-buyer-personal-profile:display'))
+
+    assert response.status_code == 200
 
 
 @mock.patch.object(api_client.company, 'verify_identity_request')