+++ profile/fab/views.py
@@ -47,23 +47,13 @@ class FindABuyerView(TemplateView):
             template_name = self.template_name_not_fab_user
         return [template_name]
 
-    def is_company_profile_owner(self):
-        if not self.request.user.company:
-            return False
-        response = api_client.supplier.retrieve_profile(self.request.user.session_id)
-        response.raise_for_status()
-        parsed = response.json()
-        return parsed['is_company_owner']
-
     def get_context_data(self):
         if self.request.user.is_authenticated and self.request.user.company:
             company = self.request.user.company.serialize_for_template()
         else:
             company = None
-
         return {
             'fab_tab_classes': 'active',
-            'is_profile_owner': self.is_company_profile_owner(),
             'company': company,
             'FAB_EDIT_COMPANY_LOGO_URL': settings.FAB_EDIT_COMPANY_LOGO_URL,
             'FAB_EDIT_PROFILE_URL': settings.FAB_EDIT_PROFILE_URL,
@@ -627,3 +617,26 @@ class IdentityVerificationRequestFormView(SuccessMessageMixin, FormView):
         response = api_client.company.verify_identity_request(self.request.user.session_id)
         response.raise_for_status()
         return super().form_valid(form)
+
+
+class PersonalProfileEditFormView(core.mixins.CreateUserProfileMixin, SuccessMessageMixin, FormView):
+    template_name = 'fab/personal-profile-edit-form.html'
+    form_class = forms.PersonalProfileEdit
+    success_url = reverse_lazy('find-a-buyer-personal-profile:display')
+    success_message = 'Personal details updated'
+
+    def get_initial(self):
+        return {
+            'given_name': self.request.user.first_name,
+            'family_name': self.request.user.last_name,
+            'job_title': self.request.user.job_title,
+            'phone_number': self.request.user.mobile_phone_number,
+        }
+
+    def form_valid(self, form):
+        self.update_user_profile(form)
+        return super().form_valid(form)
+
+
+class PersonalProfileView(TemplateView):
+    template_name = 'fab/personal-profile.html'