+++ sso/models.py
@@ -9,6 +9,8 @@ from profile.fab import helpers
 
 class SSOUser(directory_sso_api_client.models.SSOUser):
     has_user_profile = models.BooleanField()
+    job_title = models.CharField(max_length=123)
+    mobile_phone_number = models.CharField(max_length=128)
 
     @cached_property
     def company(self):
@@ -22,6 +24,8 @@ class SSOUser(directory_sso_api_client.models.SSOUser):
 
     @property
     def is_company_admin(self):
+        if not self.supplier:
+            return False
         return self.supplier['role'] == user_roles.ADMIN
 
     @property